rename steps directory to stepper
