+++ src/EditModel/event-program/EventProgramStepperContent.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import mapPropsStream from 'recompose/mapPropsStream';
 import { get, compose } from 'lodash/fp';
-import { createStepperContentFromConfig } from '../steps/stepper';
+import { createStepperContentFromConfig } from '../stepper/stepper';
 import { activeStepSelector } from './selectors';
 import eventProgramStore from './eventProgramStore';
 import steps from './event-program-steps';
