+++ src/EditModel/event-program/notifications/NotificationDialog.js
@@ -2,7 +2,7 @@ import React, { PropTypes, Component } from 'react';
 import { connect } from 'react-redux';
 import FlatButton from 'material-ui/FlatButton';
 import Dialog from 'material-ui/Dialog';
-import { createStepperFromConfig } from '../../steps/stepper';
+import { createStepperFromConfig } from '../../stepper/stepper';
 import withState from 'recompose/withState';
 import compose from 'recompose/compose';
 import withProps from 'recompose/withProps';
