+++ src/EditModel/program-indicator/ProgramIndicatorStepper.js
@@ -1,4 +1,4 @@
-import { createStepperFromConfig } from '../steps/stepper';
+import { createStepperFromConfig } from '../stepper/stepper';
 import { activeStepSelector } from './selectors';
 import { connect } from 'react-redux';
 import steps from './program-indicator-steps';
